+++ changelog.md
@@ -1,6 +1,18 @@
-# 11.1.0
+# 12.0.0
 
-Saves from 11.0.0 are compatible with 11.1.0. See Known Issues section for exceptions.
+Saves from 11.x are not compatible with 12.0.0.
+
+## Features/Improvements
+
+* **[Campaign]** Removed deprecated settings for generating persistent and invulnerable AWACs and tankers.
+* **[Mods]** F/A-18 E/F/G Super Hornet mod version updated to 2.3.
+
+## Fixes
+
+
+# 11.1.1
+
+Saves from 11.0.0 are compatible with 11.1.1. See Known Issues section for exceptions.
 
 ## Features/Improvements